+++ core-web/libs/block-editor/src/lib/extensions/bubble-form/bubble-form.component.html
@@ -1,14 +1,17 @@
 <form *ngIf="form" [formGroup]="form" (ngSubmit)="onSubmit()">
     <div *ngFor="let control of dynamicControls" class="form-row">
         <label [for]="control.key">{{ control.label | titlecase }}</label>
-        <input
-            #group
-            pInputText
-            type="control.type"
-            [formControlName]="control.key"
-            [id]="control.key"
-            [style]="{ width: '100%', fontSize: '14px', height: '40px' }"
-        />
+        <ng-container [ngSwitch]="control.type">
+            <input
+                *ngSwitchDefault
+                #group
+                pInputText
+                type="control.type"
+                [formControlName]="control.key"
+                [id]="control.key"
+                [style]="{ width: '100%', fontSize: '14px', height: '40px' }"
+            />
+        </ng-container>
     </div>
 
     <div class="form-action">
