+++ core-web/libs/block-editor/src/lib/extensions/bubble-form/bubble-form.component.ts
@@ -2,21 +2,12 @@ import {
     Component,
     EventEmitter,
     Output,
-    Input,
     ElementRef,
     ViewChildren,
     QueryList
 } from '@angular/core';
 import { FormBuilder, FormGroup } from '@angular/forms';
-
-export interface DynamicControl<T> {
-    value?: T;
-    key?: string;
-    label?: string;
-    required?: boolean;
-    controlType?: string;
-    type?: string;
-}
+import { DynamicControl } from './model';
 
 @Component({
     selector: 'dot-bubble-form',
@@ -25,9 +16,11 @@ export interface DynamicControl<T> {
 })
 export class BubbleFormComponent {
     @ViewChildren('group') inputs: QueryList<ElementRef>;
+
     @Output() formValues = new EventEmitter();
     @Output() hide = new EventEmitter<boolean>();
-    @Input() dynamicControls: DynamicControl<string>[] = [];
+
+    dynamicControls: DynamicControl<unknown>[] = [];
     form: FormGroup;
 
     constructor(private fb: FormBuilder) {}
@@ -40,7 +33,8 @@ export class BubbleFormComponent {
         this.form.setValue(values);
     }
 
-    buildForm() {
+    buildForm(controls: DynamicControl<unknown>[]) {
+        this.dynamicControls = controls;
         this.form = this.fb.group({});
         this.dynamicControls.forEach((control) => {
             this.form.addControl(control.key, this.fb.control(control.value || null));
