+++ core-web/libs/block-editor/src/lib/extensions/bubble-form/model/index.ts
@@ -1,0 +1,8 @@
+export interface DynamicControl<T> {
+    value?: T;
+    key?: string;
+    label?: string;
+    required?: boolean;
+    controlType?: string;
+    type?: string;
+}
