+++ core-web/libs/block-editor/src/lib/extensions/bubble-form/utils/index.ts
@@ -1,0 +1,23 @@
+import { DynamicControl } from '../model';
+
+export const imageFormControls: DynamicControl<string>[] = [
+    {
+        key: 'src',
+        label: 'path',
+        required: true,
+        controlType: 'text',
+        type: 'text'
+    },
+    {
+        key: 'alt',
+        label: 'alt',
+        controlType: 'text',
+        type: 'text'
+    },
+    {
+        key: 'title',
+        label: 'caption',
+        controlType: 'text',
+        type: 'text'
+    }
+];
